Lesson 24 Hibernate
- create repository implementations with hibernate

+++ src/main/java/com/alevel/lesson10/shop/repository/InvoiceRepository.java
@@ -18,5 +18,5 @@ public interface InvoiceRepository {
 
     long getInvoiceCount();
 
-    Map<Long, Long> groupBySum();
+    Map<Double, Long> groupBySum();
 }
